Fixed the themes preview error.
On the common site visit, the `current_theme` is instantiated using `current_site.get_theme.decorate` method chain.

If the FrontendController receives the `ccc_theme_preview` param, the current_theme is instantiated from this param's value, but it is not used, because the `current_site.get_theme_slug` is being appended to the `lookup_context.prefixes`.

So, given the correct behaviour of different methods of `current_theme` instantiation, the fix was to append `current_theme.name` instead of `current_site.get_theme_slug`  to the `lookup_context.prefixes`.